pyproject.toml
[project]
name = "ligma"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.7.1",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.2",
]
